Changed how the pools for special panels work
There are a number of panel pools intended just to make sure that a specific panel placement receives an appropriate swap (Quarry Laser, Town Windmill Control, Mill Elevator Control, Symmetry Door 2, and UTM Perspective 4). Previously, every panel within each pool was shuffled amongst themselves, which, combined with the fact that some panels were in multiple pools, could cause a butterfly effect where some panels ended up swapped with something that they did not share a pool with. This change makes it so that each special panel is swapped with a single other panel from its respective pool, and then both the special panel and the panel it was swapped with are no longer available for swapping.

quarryLaserOptions was inverted to being quarryLaserBanned because it is much smaller as an exclusion list.

Double randomizer mode no longer shuffles Symmetry Door 2, because I'm not convinced that the current pool of panels there are actually guaranteed to pass through the latches after applying Sigma's randomiser.