Expanded pool of panels for Symmetry Island Door 2
I manually checked every item in squarePanels and came up with a list of panels that are banned in that location because they don't require the latches to be open. Every other panel is a viable option.

Double mode currently does not randomise this panel however because I'm not sure if every panel in the pool is still viable after using Sigma's randomiser. This also raises the question of whether or not the Quarry Laser banlist is appropriate for Sigma's too.